Put the tx related helper functions in a separate file.
