+++ .gitignore
@@ -1,3 +1,4 @@
+logs.txt
 .DS_Store
 node_modules
 results
