+++ espresso/environment/7_stateless_batcher_test.go
@@ -6,8 +6,10 @@ import (
 	"github.com/ethereum-optimism/optimism/op-e2e/system/e2esys"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"math/big"
 	"testing"
+	"time"
 )
 
 // TestStatelessBatcher is a test that verifies a batcher can operate (especially restart) correctly and efficiently without persistent storage.
@@ -57,7 +59,7 @@ func TestStatelessBatcher(t *testing.T) {
 	caffVerif := system.NodeClient(env.RoleCaffNode)
 
 	// Fund Alice
-	runSimpleL1TransferAndVerifier(ctx, t, system)
+	env.RunSimpleL1TransferAndVerifier(ctx, t, system)
 
 	balanceAliceInitial, err := l2Verif.BalanceAt(ctx, addressAlice, nil)
 	if have, want := err, error(nil); have != want {
@@ -77,35 +79,36 @@ func TestStatelessBatcher(t *testing.T) {
 
 	var caffBalanceNew *big.Int
 
-	//driver := system.BatchSubmitter.TestDriver()
+	driver := system.BatchSubmitter.TestDriver()
 	numIterations := 10
 
 	// We select a range of iterations when the batcher is turned off.
-	//turnBatcherOffIteration := rand.IntN(numIterations / 2)
-	//turnBatcherOnIteration := rand.IntN(numIterations/2) + numIterations/2
+	turnBatcherOffIteration := 10 //rand.IntN(numIterations / 2)
+	turnBatcherOnIteration := 10  // rand.IntN(numIterations/2) + numIterations/2
 
 	batcherIsUp := true
 	for i := 0; i < numIterations; i++ {
 
 		t.Log("******************* Iteration: ", i)
 		//Let us stop the batcher
-		//if i == turnBatcherOffIteration {
-		//	err = driver.StopBatchSubmitting(ctx)
-		//	require.NoError(t, err)
-		//	time.Sleep(2 * time.Second)
-		//	batcherIsUp = false
-		//}
-		//
+		if i == turnBatcherOffIteration {
+			err = driver.StopBatchSubmitting(ctx)
+			require.NoError(t, err)
+			time.Sleep(2 * time.Second)
+			batcherIsUp = false
+		}
+
 		//// Let us start the batcher again
-		//if i == turnBatcherOnIteration {
-		//	err = driver.StartBatchSubmitting()
-		//	require.NoError(t, err)
-		//	batcherIsUp = true
-		//}
+		if i == turnBatcherOnIteration {
+			err = driver.StartBatchSubmitting()
+			require.NoError(t, err)
+			batcherIsUp = true
+		}
 
 		// The batcher is up, we can send coins
 		if batcherIsUp {
-			runSimpleL2Transfer(ctx, t, system, uint64(i+1), *amount, l2Seq, l2Verif)
+			// Nonce is i+1 because nonce 0 is used for the initial deposit
+			env.RunSimpleL2Transfer(ctx, t, system, uint64(i+1), *amount, l2Seq, l2Verif)
 			numDeposits++
 		}
